requirements.txt
djangorestframework
django-redis-cache
django-debug-toolbar
python-dotenv
psycopg2-binary
whitenoise
gunicorn
rq
PyGithub
boto3
reliability
matplotlib
scipy
django-rq